requirements.txt
streamlit
groq
frontend
python-dotenv
sentencepiece
transformers
loguru
PyMuPDF
sentence-transformers
fpdf
openai